use double-casted values of 1e99/1e98 for SCIP_INVALID/SCIP_UNKNOWN
- when running on 32bit, floating point values larger 1e22 cannot be
  represented exactly as double, thus code like
    double a = 1e99;
    assert(a==1e99);
  raises an assert
- however, assert(a==(double)1e99); would succeed
- thus, make sure that SCIP_INVALID, etc., are defined to
  double-representable values
- if replacing -std=c99 by -std=gnu99 in the compiler flags, this goes
  away
- thanks to @bzfwinkm to figure this out

+++ src/scip/def.h
@@ -153,8 +153,8 @@ extern "C" {
 #define SCIP_DEFAULT_HUGEVAL          1e+15  /**< values larger than this are considered huge and should be handled separately (e.g., in activity computation) */
 #define SCIP_MAXEPSILON               1e-03  /**< maximum value for any numerical epsilon */
 #define SCIP_MINEPSILON               1e-20  /**< minimum value for any numerical epsilon */
-#define SCIP_INVALID                  1e+99  /**< floating point value is not valid */
-#define SCIP_UNKNOWN                  1e+98  /**< floating point value is not known (in primal solution) */
+#define SCIP_INVALID          (double)1e+99  /**< floating point value is not valid */
+#define SCIP_UNKNOWN          (double)1e+98  /**< floating point value is not known (in primal solution) */
 
 
 #define REALABS(x)        (fabs(x))
